+++ babel.config.json
@@ -1,3 +1,7 @@
 {
-    "presets": ["@babel/preset-env"]
+  "env": {
+    "test": {
+      "plugins": ["@babel/plugin-transform-modules-commonjs"]
+    }
   }
+}
